requirements.txt
numpy==1.19.5
pandas==1.4.0
parse==1.19.0
pydicom==2.1.2
pymongo==4.0.1
pytest==6.2.5
PyYAML==6.0
